internal CloseCommand
