+++ MahApps.Metro/Controls/MetroTabItem.cs
@@ -1,49 +1,9 @@
-﻿using System;
-using System.Windows;
+﻿using System.Windows;
 using System.Windows.Controls;
 using System.Windows.Input;
 
 namespace MahApps.Metro.Controls
 {
-    internal class MetroTabItemCloseCommand : ICommand
-    {
-        private readonly Func<object, bool> canExecute;
-        private readonly Action<object> executeAction;
-
-        public MetroTabItemCloseCommand(Func<object, bool> canExecute, Action<object> executeAction)
-        {
-            this.canExecute = canExecute;
-            this.executeAction = executeAction;
-        }
-
-        /// <summary>
-        /// Defines the method that determines whether the command can execute in its current state.
-        /// </summary>
-        /// <returns>
-        /// true if this command can be executed; otherwise, false.
-        /// </returns>
-        /// <param name="parameter">Data used by the command.  If the command does not require data to be passed, this object can be set to null.</param>
-        public bool CanExecute(object parameter)
-        {
-            return this.canExecute != null && this.canExecute(parameter);
-        }
-
-        /// <summary>
-        /// Defines the method to be called when the command is invoked.
-        /// </summary>
-        /// <param name="parameter">Data used by the command.  If the command does not require data to be passed, this object can be set to null.</param>
-        public void Execute(object parameter)
-        {
-            this.executeAction?.Invoke(parameter);
-        }
-
-        public event EventHandler CanExecuteChanged
-        {
-            add { CommandManager.RequerySuggested += value; }
-            remove { CommandManager.RequerySuggested -= value; }
-        }
-    }
-
     /// <summary>
     /// An extended TabItem with a metro style.
     /// </summary>
@@ -52,7 +12,7 @@ namespace MahApps.Metro.Controls
         public MetroTabItem()
         {
             DefaultStyleKey = typeof(MetroTabItem);
-            this.InternalCloseTabCommand = new MetroTabItemCloseCommand(InternalCloseTabCommandCanExecute, InternalCloseTabCommandExecuteAction);
+            this.InternalCloseTabCommand = new CloseCommand(InternalCloseTabCommandCanExecute, InternalCloseTabCommandExecuteAction);
         }
 
         private void InternalCloseTabCommandExecuteAction(object o)
